requirements.txt
### **requirements.txt**
```txt
pandas==2.0.3
numpy==1.24.3
sentence-transformers==2.2.2
scann==1.3.1
torch==2.0.1
scikit-learn==1.3.0
pyyaml==6.0
python-dotenv==1.0.0
openpyxl==3.1.2
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.65.0
pytest==7.4.0
black==23.7.0
flake8==6.0.0